Testing
Testing a few connections to the AWS server without sucess. Hosted the website and got errors due to the connection failing.

Added some more data to the servers, need to figure out how to connect to a AWS server. Hopefully can correctly set up an apache server/php/myphpadmin to be able to test by myself.